third_party/blink/web_tests/platform/mac/external/wpt/css/cssom-view/scrollIntoView-stuck.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] CSSOM View - scrollIntoView doesn't consider scroll-padding when target is stuck
  assert_equals: Shouldn't have scrolled expected 0 but got 2421
Harness: the test ran to completion.
